Remove superfluous tutorial step
