+++ Sources/Player/Player.docc/Tutorials/creating-basic-user-interface/creating-basic-user-interface-1-1.swift
@@ -1,5 +1,8 @@
+import PillarboxPlayer
 import SwiftUI
 
 struct ContentView: View {
-    
+    @StateObject private var player = Player(
+        item: .simple(url: URL(string: "https://devstreaming-cdn.apple.com/videos/streaming/examples/img_bipbop_adv_example_ts/master.m3u8")!)
+    )
 }
