+++ Sources/Player/Player.docc/Tutorials/creating-basic-user-interface/creating-basic-user-interface-1-2.swift
@@ -5,4 +5,8 @@ struct ContentView: View {
     @StateObject private var player = Player(
         item: .simple(url: URL(string: "https://devstreaming-cdn.apple.com/videos/streaming/examples/img_bipbop_adv_example_ts/master.m3u8")!)
     )
+
+    var body: some View {
+        VideoView(player: player)
+    }
 }
