+++ Sources/Player/Player.docc/Tutorials/creating-basic-user-interface/creating-basic-user-interface-1-3.swift
@@ -7,6 +7,18 @@ struct ContentView: View {
     )
 
     var body: some View {
-        VideoView(player: player)
+        ZStack {
+            VideoView(player: player)
+            playbackButton()
+        }
+    }
+
+    private func playbackButton() -> some View {
+        Button(action: player.togglePlayPause) {
+            Image(systemName: "play.circle.fill")
+                .resizable()
+                .aspectRatio(contentMode: .fit)
+                .frame(width: 50)
+        }
     }
 }
