+++ Sources/Player/Player.docc/Tutorials/creating-basic-user-interface/creating-basic-user-interface-1-4.swift
@@ -15,7 +15,7 @@ struct ContentView: View {
 
     private func playbackButton() -> some View {
         Button(action: player.togglePlayPause) {
-            Image(systemName: "play.circle.fill")
+            Image(systemName: player.shouldPlay ? "pause.circle.fill" : "play.circle.fill")
                 .resizable()
                 .aspectRatio(contentMode: .fit)
                 .frame(width: 50)
